requirements.txt
click==8.1.7
Flask==3.0.0
gunicorn==21.2.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
Werkzeug==3.0.1
email-validator==2.0.0
Flask-Mail==0.9.1
Flask-WTF==1.2.1